+++ src/packages/babel.js
@@ -1,54 +1,54 @@
 module.exports = {
-  "base": [
+  base: [
     {
-      "@babel/cli": {
-        "version": "7.15.7",
-        "type": "--save-dev"
-      }
+      '@babel/cli': {
+        version: '7.15.7',
+        type: '--save-dev',
+      },
     },
     {
-      "@babel/core": {
-        "version": "7.15.5",
-        "type": "--save-dev"
-      }
+      '@babel/core': {
+        version: '7.15.5',
+        type: '--save-dev',
+      },
     },
     {
-      "@babel/preset-env": {
-        "version": "7.15.6",
-        "type": "--save-dev"
-      }
+      '@babel/preset-env': {
+        version: '7.15.6',
+        type: '--save-dev',
+      },
     },
     {
-      "@babel/runtime": {
-        "version": "7.15.4",
-        "type": "--save-dev"
-      }
+      '@babel/runtime': {
+        version: '7.15.4',
+        type: '--save-dev',
+      },
     },
     {
-      "@babel/plugin-proposal-throw-expressions": {
-        "version": "7.14.5",
-        "type": "--save-dev"
-      }
+      '@babel/plugin-proposal-throw-expressions': {
+        version: '7.14.5',
+        type: '--save-dev',
+      },
     },
     {
-      "@babel/plugin-transform-arrow-functions": {
-        "version": "7.14.5",
-        "type": "--save-dev"
-      }
+      '@babel/plugin-transform-arrow-functions': {
+        version: '7.14.5',
+        type: '--save-dev',
+      },
     },
     {
-      "babel-loader": {
-        "version": "8.2.2",
-        "type": "--save-dev"
-      }
-    }
+      'babel-loader': {
+        version: '8.2.2',
+        type: '--save-dev',
+      },
+    },
   ],
-  "react": [
+  react: [
     {
-      "@babel/preset-react": {
-        "version": "7.14.5",
-        "type": "--save-dev"
-      }
-    }
-  ]
+      '@babel/preset-react': {
+        version: '7.14.5',
+        type: '--save-dev',
+      },
+    },
+  ],
 }