+++ src/packages/esProjectMaker.js
@@ -2,7 +2,7 @@ module.exports = {
   base: [
     {
       '@mparulski/es-project-maker': {
-        version: '0.16.0',
+        version: '0.17.0',
         type: '--save-dev',
       },
     },