+++ src/packages/webpack.js
@@ -20,7 +20,7 @@ module.exports = {
     },
     {
       'webpack-dev-server': {
-        version: '4.2.1',
+        version: '4.3.0',
         type: '--save-dev',
       },
     },